+++ mobile/android/android-components/docs/changelog.md
@@ -15,6 +15,9 @@ permalink: /changelog/
 * **feature-intent**
    * ⚠️ **This is a breaking change**: `IntentProcessor.process` is not a suspend function anymore.
 
+* **feature-search**
+  * Adds optional `parentSession` to attach to a new search session in `SearchUseCases`
+
 * **feature-contextmenu**
   * Add "Share image" to context menu.
 
@@ -49,8 +52,7 @@ permalink: /changelog/
     * BUGFIX: Compile dependencies with `NDEBUG` to avoid linking unavailable symbols.
       This fixes a crash due to a missing `stderr` symbol on older Android.
 
-* **feature-webnotifications**
-  * `WebNotificationFeature` checks the site permissions first before showing a notification.
+
 
 # 47.0.0